commit lo

+++ DARI_TN/src/main/java/tn/dari/spring/repository/BankRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import tn.dari.spring.entity.Bank;
 
-public interface BankRepository extends JpaRepository<Bank, Long>{
+public interface BankRepository extends JpaRepository<Bank, Integer>{
 
 }
